App2/requirements.txt
tensorflow-cpu == 2.6.0
flask
numpy == 1.19.5
pandas
sklearn
gunicorn
regex
Sastrawi